refactor: 묵찌빠, 가위바위보 구조체 분리
