+++ RockPaperScissors/RockPaperScissors/MookJjiPpa.swift
@@ -2,7 +2,7 @@
 //  MookJjiPpa.swift
 //  RockPaperScissors
 //
-//  Created by Kiwon Song on 2022/04/20.
+//  Created by kiwi,bard on 2022/04/20.
 //
 
 enum MookJjiPpa: Int {
